Online Shop Template Integration
Added templates for the website to have the base. Later on, will start adding functionality to make the website responsive and look like actual e-commerce website with products in grid in the center section with images which can be added to cart, product category on the left, search products by text box on top, pagination support for products on bottom incase if there are lot of products in the inventory and lastly shopping cart support on the top right, checking out and so on.