Add support for the Product Grid
Changed some minor code format for the logo in index.html to make it look pretty with Prettier code formatter extension in VS code.

Display the products in the grid format on the homepage. That's basically information being displayed in a grid fashion based on the bootstrap style for the container fluid in different rows and it will simply build and wrap accordingly based on the length of the product.