Enhancing Application with Product Images
Run the db script to refresh db with 100 products. By default, Spring Data REST only returns the first page of 20 items. Did a quick workaround to modify it to tell Spring Data REST to return specific number of items i.e. 100 (all products) for us to test it and it worked; reverted it back to the default one.