Define the Router Outlet and Set up Router links to pass category id param
Defined the routes in app.module.ts

Set up Route links to app.component.html in the menu bar by passing the category followed by a back slash and param name which is "1" for the books and so on foor the menu links based on the Routes configuration where we are passing an id to the category to set it up. At this moment, hard-coded the categories but later will make it dynamic and read the information from REST API. But for now, let's get it working with basics.

Commented out the Mouse Pads and Luggage tags for now,