Enhance ProductListComponent to read category id param
Added comments for route matches. Check to see if product category is still available if yes, then convert the string it to the number using the "+" symbol and "!" at the end to make it not null when you define. If not then make it default to 1. After getting the value, pass it to the getProductList method as parameter inside the productList component.