Add Dynamic Code to HTML (Dynamic Search Component)
In HTML, replaced the hard-coded links with menu component. Did it by looping over categories and building links dynamically. I had a typo doing this to fetch the product category from API, so no categories were displayed but fixed it eventually.

Before doing this, tested with hard-coded still to use in menu component to make sure it still works.