Display Category Name in Product List Grid
The approach I took was to update the menu component. I passed the category name along with category id in the menu router link. Then in the product-list component, I can read the category name. Then I can bind this in the HTML file. This will give us this output with category name. I added a horizontal rule <hr> too to make it look better.