Send data to search
Updated search-component.html to have a new div with a search box and logic nowSearch method to do it with the user's value.

Also updated search.component.ts to inject the router into the component, getting the logs of values for debugging if needed and calling the route search with the given keyword handled by the ProductListComponent to reuse the logic and view for listing products.