Install ng-bootstrap
Was getting npm errors for trying to install it as I have Angular >= 9.0 so, simply used Angular schematics to add ng-bootstrap library to it by running the cmd
ng add @ng-bootstrap/ng-bootstrap
but as I said had npm errors.

But found the solution, for some reason running this command before adding the ng bootstrap fixed it.
npm config set legacy-peer-deps true. It automatically implemented the import in app.module.ts to expose the exported declarations in NgbModule (classes, interfaces, constants etc) and makes them available in the current module.