Checkout Form -  Populate Countries and States - Bug Fix
Just removed <div> tag, as it was of no use for checkout button.

In previous commit, as I tried to populate the billing address states with the country selected, but the code wasn't in sync or set up to have that data so, I fixed it updating the array in the component with the appropriate data by making an assignment for billingAddressStates and copy over or assign values of shippingAddressStates to make it work by populating of the dropdown list options. Unfortunately, states list still doesn't works now. so will try to fix it later.